Revert "feat: new page structure"
