+++ pages/_app.tsx
@@ -1,24 +1,17 @@
+import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
+import Theme from '../Theme'
+import Ellipse from '../components/Ellipse'
+import Layout from '../components/Layout'
 import { Provider } from 'jotai'
+import Footer from '../components/Footer'
+import { useRouter } from 'next/router'
 import Script from 'next/script'
 import CookieConsent from 'react-cookie-consent'
-import { NextPage } from 'next'
-import { ReactElement, ReactNode } from 'react'
 
-import '../styles/globals.css'
-import Theme from '../Theme'
-
-type NextPageWithLayout = NextPage & {
-  getLayout?: (page: ReactElement) => ReactNode
-}
-
-type AppPropsWithLayout = AppProps & {
-  Component: NextPageWithLayout
-}
-
-function MyApp({ Component, pageProps }: AppPropsWithLayout) {
-  const getLayout = Component.getLayout ?? ((page) => page)
+function MyApp({ Component, pageProps }: AppProps) {
+  const router = useRouter()
   return (
     <Provider>
       <Script
@@ -83,7 +76,11 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
           expires={150}>
           Denna site använder cookies för att förbättra användarupplevelsen.
         </CookieConsent>
-        {getLayout(<Component {...pageProps} />)}
+        <Ellipse />
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+        {router.pathname.indexOf('/kommun') ? <Footer /> : null}
       </Theme>
     </Provider>
   )
