--- pages/partier/index.tsx
@@ -1,65 +1,0 @@
-import { GetServerSideProps } from 'next'
-import { ReactElement } from 'react'
-import styled from 'styled-components'
-import Link from 'next/link'
-
-import MetaTags from '../../components/MetaTags'
-import { Paragraph } from '../../components/Typography'
-import { EmissionService } from '../../utils/emissionService'
-import PageWrapper from '../../components/PageWrapper'
-import Layout from '../../components/Layout'
-import Footer from '../../components/Footer'
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-`
-
-const Partier = () => {
-  return (
-    <>
-      <MetaTags
-        title="Klimatkollen — Få koll på Sveriges klimatomställning"
-        description=""
-      />
-      <PageWrapper backgroundColor="black">
-        <Container>
-          <div>
-            <Paragraph>Här kommer grafer för partierna finnas.</Paragraph>
-          </div>
-          <div>
-            <Link href="/kommuner">
-              <a href="placeholder">Kommuner</a>
-            </Link>
-          </div>
-        </Container>
-      </PageWrapper>
-    </>
-  )
-}
-
-export const getServerSideProps: GetServerSideProps = async ({ res }) => {
-  const municipalities = await new EmissionService().getMunicipalities()
-  if (municipalities.length < 1) throw new Error('No municipalities found')
-
-  res.setHeader(
-    'Cache-Control',
-    'public, stale-while-revalidate=60, max-age=' + 60 * 60 * 24 * 7,
-  )
-
-  return {
-    props: { municipalities },
-  }
-}
-
-export default Partier
-
-Partier.getLayout = function getLayout(page: ReactElement) {
-  return (
-    <>
-      <Layout>{page}</Layout>
-      <Footer />
-    </>
-  )
-}
